Create the Group model and table
This commit adds the Group model which will act as the main model for representing a group making use of talks night. This currently is owned by a single owner who will have modification rights but will likely need to be further extended to allow multiple admins at a later point in time.