Refine the space creation flow
This commit refines the space creation flow by adding the ability for user signup to redirect back to a specific path (in this case the new group path).

It also adds a feature spec which covers the creation of a group from the homepage through to being present on the group page.