Add a Talk model and talks table
This commit adds a table that can be used to contain top level information about a talk. This should just be very basic information that would be relevant to giving this talk at multiple different events. For any information that is related to a specific talks night that should be handled in a follow up model which links the group and the talk.