Add talk_suggestions table and TalkSuggestion model
This commit adds a model that represents the link between a talk and a group when the speaker has suggested their talk would be a good fit for a particular group. It also includes a specific attribute #speaker_contact which is intended to be utilised for the speaker to tell the organiser how they would like the organiser to make contact with them to discuss the talk proposal.