Add page for displaying all events for a single group
This commit adds the groups / events page which displays all previous and upcoming events for a single group.
It extracts the UI for an event summary card to a partial to be shared between both the main group page and the events list.
It adds new associations adns copes to the Event and Group model for finding previous and upcoming events.